requirements.txt
Django==2.1.2
mysqlclient==1.3.13
dj-database-url==0.3.0
python-decouple==3.1
pytz==2018.5
pillow==5.2.0
gunicorn==19.9.0
dj-database-url==0.3.0
qrcode==6.0
django-qiniu-storage==2.3.1